requirements.txt
python-telegram-bot==20.7
Pillow==10.2.0
python-dotenv==1.0.1
aiohttp==3.9.3
numpy==1.26.4
google-generativeai==0.4.1
